fix: completely reset the templates on game and template reset
The templates have to be a complete reset, otherwise, when a new game is created and no templates exist (e.g they are removed completely), the game will not get a template update and the old templates will be used.

Telling the game server to reset this will ensure the templates are updated.